feat: allow host to be configured (#2)
* Allow for DirectUpload configuration

Some apps do not host their front end assets
under the same domain as the api. This allows
for clients to point the upload at the
correct server.

* Allow for DirectUpload configuration

Some apps do not host their front end assets
under the same domain as the api. This allows
for clients to point the upload at the
correct server.

* Change Naming and add Documentation

* Allow for DirectUpload configuration

Some apps do not host their front end assets
under the same domain as the api. This allows
for clients to point the upload at the
correct server.

* Mark host, port, and protocol as optional in README

* Update Changelog