fix: make handleBeginUpload call onSuccess with [] if no files chosen
This is so it can be used as a "onSubmit" handler in a case where you’ve 
got a form that accepts uploads and includes their IDs with the 
submission of the rest of the form. It needs to actually act as submit 
even if no attachments are given.

Of course, that doesn’t make sense for the one-click attach form that 
`ActiveStorageProvider` is intended for, so it needs to noop if the 
callback is hit with [].

Co-Authored-By: Pearl Zeng <pearlzhuzeng@users.noreply.github.com>